refactor: switch to remix vite
Ref https://remix.run/docs/en/main/future/vite#setup-vite
