+++ apps/builder/env.d.ts
@@ -1,0 +1,2 @@
+/// <reference types="vite/client" />
+/// <reference types="@remix-run/node" />
